+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 0.1.1
 
+ * Remove Diffie-Hellman (since no RSA) #6
  * Add github actions as CI #5
  * Fix bad MTU packing causing flaky tests #4
  * Remove ciphers using RSA #3